requirements.txt
requests==2.31.0
webdriver-manager==4.0.1
selenium==4.19.0
google-generativeai==0.4.0